scripts/requirements.txt
# Documentation Pipeline Requirements
# Install: python -m venv .venv && source .venv/bin/activate && pip install -r scripts/requirements.txt

# Core dependencies
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
html2text>=2024.2

# Development tools
ruff>=0.4
pytest>=8.0

# Optional: for faster YAML parsing
# ruamel.yaml>=0.18
